Refactor language initialization to trait.
