+++ module/VuFind/src/VuFind/Bootstrapper.php
@@ -42,6 +42,8 @@ use Zend\Router\Http\RouteMatch;
  */
 class Bootstrapper
 {
+    use \VuFind\I18n\Translator\LanguageInitializerTrait;
+
     /**
      * Main VuFind configuration
      *
@@ -275,30 +277,6 @@ class Bootstrapper
     }
 
     /**
-     * Support method for initLanguage() -- look up all text domains.
-     *
-     * @return array
-     */
-    protected function getTextDomains()
-    {
-        $base = APPLICATION_PATH;
-        $local = LOCAL_OVERRIDE_DIR;
-        $languagePathParts = ["$base/languages"];
-        if (!empty($local)) {
-            $languagePathParts[] = "$local/languages";
-        }
-        $languagePathParts[] = "$base/themes/*/languages";
-
-        $domains = [];
-        foreach ($languagePathParts as $current) {
-            $places = glob($current . '/*', GLOB_ONLYDIR | GLOB_NOSORT);
-            $domains = array_merge($domains, array_map('basename', $places));
-        }
-
-        return array_unique($domains);
-    }
-
-    /**
      * Set up language handling.
      *
      * @return void
@@ -336,16 +314,8 @@ class Bootstrapper
             }
             try {
                 $translator = $sm->get(\Zend\Mvc\I18n\Translator::class);
-                $translator->setLocale($language)
-                    ->addTranslationFile('ExtendedIni', null, 'default', $language);
-                foreach ($this->getTextDomains() as $domain) {
-                    // Set up text domains using the domain name as the filename;
-                    // this will help the ExtendedIni loader dynamically locate
-                    // the appropriate files.
-                    $translator->addTranslationFile(
-                        'ExtendedIni', $domain, $domain, $language
-                    );
-                }
+                $translator->setLocale($language);
+                $this->addLanguageToTranslator($translator, $language);
             } catch (\Zend\Mvc\I18n\Exception\BadMethodCallException $e) {
                 if (!extension_loaded('intl')) {
                     throw new \Exception(
